+++ scilab/routines/graphics/periGif.c
@@ -1285,7 +1285,7 @@ void C2F(boundingboxGif)(char *string, integer *x, integer *y, integer *rect, in
     return;
   }
   width = 0;
-  for (k=0;k<strlen(string);k++) 
+  for (k=0;k< (int) strlen(string);k++) 
     width += gdCharWidth(GifFont, string[k]);
   rect[0]= (int)(*x);
   rect[1]= (int)(*y);
@@ -1528,7 +1528,7 @@ void C2F(drawarcsGif)(char *str, integer *vects, integer *style, integer *n, int
 void C2F(drawarcGif)(char *str, integer *x, integer *y, integer *width, integer *height, integer *angle1, integer *angle2, double *dv1, double *dv2, double *dv3, double *dv4)
 { 
   integer vx[365],vy[365],k,n;
-  float alpha,fact=0.01745329251994330,w,h;
+  float alpha,fact= (float) 0.01745329251994330,w,h;
   integer close = 0;
 
   if (GifIm == (gdImagePtr)0 ) {
@@ -1536,8 +1536,8 @@ void C2F(drawarcGif)(char *str, integer *x, integer *y, integer *width, integer 
     return;
   }
 
-  w = (*width)/2.0;
-  h = (*height)/2.0;
+  w = (float) ((*width)/2.0);
+  h = (float) ((*height)/2.0);
   n=Min((*angle2/64),360);
   for (k = 0; k < n; ++k) {
     alpha=((*angle1/64)+k)*fact;
@@ -1553,7 +1553,7 @@ void C2F(drawarcGif)(char *str, integer *x, integer *y, integer *width, integer 
 void C2F(fillarcGif)(char *str, integer *x, integer *y, integer *width, integer *height, integer *angle1, integer *angle2, double *dv1, double *dv2, double *dv3, double *dv4)
 { 
   integer vx[365],vy[365],k,k0,kmax,n;
-  float alpha,fact=0.01745329251994330,w,h;
+  float alpha,fact= (float) 0.01745329251994330,w,h;
   integer close = 1;
   if (GifIm == (gdImagePtr)0 ) {
     sciprint(" xinit must be called before any action \r\n");
@@ -1561,8 +1561,8 @@ void C2F(fillarcGif)(char *str, integer *x, integer *y, integer *width, integer 
   }
   n=Min((*angle2/64),360);
 
-  w = (*width)/2.0;
-  h = (*height)/2.0;
+  w = (float) ((*width)/2.0);
+  h = (float) ((*height)/2.0);
   k0 = 0;
   kmax = n-1;
